requirements.txt
-e .
aws-cdk.core==1.61.0 #CORE Module
aws-cdk.aws_s3==1.61.0